chore: moving the reflist to components
