+++ shesha-reactjs/src/components/chevron/index.tsx
@@ -4,7 +4,7 @@ import { Button, Form, FormInstance } from "antd";
 import ConfigurableButton from "@/designer-components/button/configurableButton";
 import { pickStyleFromModel, useTheme } from "@/index";
 import { useStyles } from "./styles";
-import { fadeColor } from "@/providers/refList/provider/utils";
+import { fadeColor } from "@/components/refListSelectorDisplay/provider/utils";
 import { IChevronButton, IChevronControlProps } from "./models";
 import classNames from "classnames";
 import { addPx } from '@/utils/style';
