+++ shesha-reactjs/src/components/kanban/components/renderColumn.tsx
@@ -1,7 +1,7 @@
 import { getDimensionsStyle } from '@/designer-components/_settings/utils/dimensions/utils';
 import { getFontStyle } from '@/designer-components/_settings/utils/font/utils';
 import { ConfigurableForm, ShaIcon, useAvailableConstantsData, useConfigurableActionDispatcher } from '@/index';
-import { useRefListItemGroupConfigurator } from '@/providers/refList/provider';
+import { useRefListItemGroupConfigurator } from '@/components/refListSelectorDisplay/provider';
 import { LeftOutlined, MoreOutlined, PlusOutlined, RightOutlined, SettingOutlined } from '@ant-design/icons';
 import { Button, Dropdown, Flex, MenuProps, Popconfirm } from 'antd';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
