+++ shesha-reactjs/src/components/refListSelectorDisplay/options/properties.tsx
