+++ shesha-reactjs/src/components/refListSelectorDisplay/options/refListItem.tsx
