+++ shesha-reactjs/src/components/refListSelectorDisplay/options/refListItemsSettings.json
