+++ shesha-reactjs/src/components/refListSelectorDisplay/options/refListItemsSettingsForm.ts
