+++ shesha-reactjs/src/components/refListSelectorDisplay/provider/index.tsx
@@ -4,19 +4,19 @@ import {
   RefListItemGroupConfiguratorActionsContext,
   RefListItemGroupConfiguratorStateContext,
   REF_LIST_ITEM_GROUP_CONTEXT_INITIAL_STATE,
-  IUpdateChildItemsPayload,
-} from './contexts';
+  IUpdateChildItemsPayload
+} from '@/components/refListSelectorDisplay/provider/contexts';
 import {
   selectItemAction,
   setItems,
   storeSettingsAction,
   updateChildItemsAction,
-  updateItemAction,
-} from './actions';
-import { RefListGroupItemProps } from './models';
-import { useGet } from '@/hooks';
-import RefListItemGroupReducer from './reducers';
-import { getItemById, getRefListItems } from './utils';
+  updateItemAction
+} from '@/components/refListSelectorDisplay/provider/actions';
+import { RefListGroupItemProps } from '@/components/refListSelectorDisplay/provider/models';
+import RefListItemGroupReducer from '@/components/refListSelectorDisplay/provider/reducers';
+import { getItemById } from '@/components/refListSelectorDisplay/provider/utils';
+import { useReferenceListDispatcher } from '@/providers/referenceListDispatcher';
 
 
 export interface IRefListItemGroupConfiguratorProviderPropsBase {
@@ -31,9 +31,10 @@ export interface IRefListItemGroupConfiguratorProviderProps {
   referenceList?: any;
 }
 
-const RefListItemGroupConfiguratorProvider: FC<PropsWithChildren<IRefListItemGroupConfiguratorProviderProps>> = (props) => {
+const RefListSelectorDisplayProvider: FC<PropsWithChildren<IRefListItemGroupConfiguratorProviderProps>> = (props) => {
   const { children, readOnly } = props;
-  const { refetch } = useGet({ path: '', lazy: true });
+ // const { refetch } = useGet({ path: '', lazy: true });
+  const { getReferenceList } = useReferenceListDispatcher();
 
   const [state, dispatch] = useReducer(RefListItemGroupReducer, {
     ...REF_LIST_ITEM_GROUP_CONTEXT_INITIAL_STATE,
@@ -41,18 +42,12 @@ const RefListItemGroupConfiguratorProvider: FC<PropsWithChildren<IRefListItemGro
     readOnly: readOnly,
   });
 
-
   useEffect(() => {
-    if (props?.items?.length && props.items.some(x =>x.referenceList?.id === props?.referenceList?.id )) return;
-    refetch(getRefListItems(props.referenceList?.id as string))
-    .then((resp) => {
-      dispatch(setItems(resp.result.items));
-    })
-    .catch((_e) => {
-      console.error("LOG:: reference list error", _e);
+    if (props?.items?.length && props.items.some(x =>x.referenceList === props?.referenceList )) return;  
+    getReferenceList({ refListId: { module: "Shesha", name:props?.referenceList?._displayName } }).promise.then((t)=>{
+      dispatch(setItems(t?.items));
     });
-
-  }, [props?.referenceList?.id]);
+  }, [props?.referenceList]);
 
   const selectItem = (uid: string) => {
     dispatch(selectItemAction(uid));
@@ -119,4 +114,4 @@ function useRefListItemGroupConfigurator() {
   return { ...useRefListItemGroupConfiguratorState(), ...useRefListItemGroupConfiguratorActions() };
 }
 
-export { RefListItemGroupConfiguratorProvider, useRefListItemGroupConfigurator };
+export { RefListSelectorDisplayProvider as RefListItemGroupConfiguratorProvider, useRefListItemGroupConfigurator };
