+++ shesha-reactjs/src/designer-components/chevron/chevron.tsx
@@ -3,7 +3,7 @@ import { FolderOpenOutlined } from '@ant-design/icons';
 import React from 'react';
 import { IToolboxComponent } from '@/interfaces';
 import { ChevronControl } from '@/components/chevron';
-import { RefListItemGroupConfiguratorProvider } from '@/providers/refList/provider';
+import { RefListItemGroupConfiguratorProvider } from '@/components/refListSelectorDisplay/provider';
 import { getSettings } from './settingsForm';
 import { ConfigurableFormItem } from '@/components';
 import { IChevronProps } from '@/components/chevron/models';
