+++ shesha-reactjs/src/designer-components/inputComponent/index.tsx
@@ -28,7 +28,7 @@ import { ItemListConfiguratorModal } from '../itemListConfigurator/itemListConfi
 import { ConfigurableActionConfigurator } from '../configurableActionsConfigurator/configurator';
 import { formTypes } from '../entityReference/settings';
 import { SortingEditor } from '@/components/dataTable/sortingConfigurator';
-import RefListItemSelectorSettingsModal from '@/providers/refList/options/modal';
+import RefListItemSelectorSettingsModal from '@/components/refListSelectorDisplay/options/modal';
 import { FormLayout } from 'antd/es/form/Form';
 import { editModes, getEditor, iconElement } from './utils';
